fix: Resolve all clippy warnings across codebase
- Add #[allow(clippy::assertions_on_constants)] to compile-time validation tests
- Remove empty lines after doc comments
- Replace range loops with iterator patterns (enumerate, iter_mut)
- Use std::slice::from_ref instead of &[x.clone()]
- Remove duplicated #![allow(dead_code)] attribute
- Fix unnecessary parentheses and format! in println!
- Change vec![] to array literal where appropriate
- Replace PI/E approximations (3.14, 2.71) with non-constant values in tests
- Add #![allow(clippy::approx_constant)] to math function test files

+++ crates/vibesql-executor/tests/join_memory_test.rs
@@ -5,7 +5,6 @@ use vibesql_executor::SelectExecutor;
 use vibesql_parser::Parser;
 use vibesql_storage::{Database, Row};
 use vibesql_types::{DataType, SqlValue};
-use vibesql_ast;
 
 /// Helper function to parse SELECT statements
 fn parse_select(sql: &str) -> vibesql_ast::SelectStmt {
@@ -59,7 +58,7 @@ fn setup_test_db() -> Database {
             SqlValue::Integer(i),
             SqlValue::Integer(custkey),
             SqlValue::Varchar(date.to_string()),
-            SqlValue::Integer((i % 3) as i64),
+            SqlValue::Integer(i % 3),
         ]);
         db.insert_row("ORDERS", row).unwrap();
     }
